payment/migrations/models/12_20240111184427_None.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        CREATE TABLE IF NOT EXISTS "cards" (
    "id" UUID NOT NULL  PRIMARY KEY,
    "created" TIMESTAMPTZ NOT NULL  DEFAULT CURRENT_TIMESTAMP,
    "updated" TIMESTAMPTZ NOT NULL  DEFAULT CURRENT_TIMESTAMP,
    "user" UUID NOT NULL,
    "number" VARCHAR(128) NOT NULL,
    "expiration_date" DATE NOT NULL,
    "owner" VARCHAR(128) NOT NULL,
    "balance" DECIMAL(10,2) NOT NULL  DEFAULT 0
);
COMMENT ON TABLE "cards" IS 'Card user model';
CREATE TABLE IF NOT EXISTS "aerich" (
    "id" SERIAL NOT NULL PRIMARY KEY,
    "version" VARCHAR(255) NOT NULL,
    "app" VARCHAR(100) NOT NULL,
    "content" JSONB NOT NULL
);"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        """
